auth-service/requirements.txt
Flask==2.2.5
Flask-Classful==0.14.1
Flask-Cors==3.0.10
Flask-JSGlue==0.3.1
Flask-KVSession-fork==0.6.4
Flask-Limiter==2.6.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==4.1.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Flask-PyMongo==2.3.0
pymongo==4.3.3
python-dotenv==1.0.0
cryptography==41.0.4 
flask-bcrypt==1.0.1
PyJWT==2.8.0